Added a handler for the group list server message. Split the packet handlers into client and server specific ones.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/gsoc_bzauthd@19987 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzAuthCommon/Socket.h
@@ -118,7 +118,7 @@ public:
 
   PlaceHolder append_placeholder(size_t size)
   {
-    int wpoz = m_wpoz;
+    size_t wpoz = m_wpoz;
     ensure_wpoz_inc(size);
     m_wpoz += size;
     return PlaceHolder(*this, size, wpoz);
